[multi forge] start multi forge, and stop current forge
