+++ src/cli/node/install/install.js
@@ -20,7 +20,7 @@ const {
   RELEASE_ASSETS,
   DEFAULT_MIRROR,
 } = require('core/env');
-const { findServicePid } = require('forge-process');
+const { findServicePid } = require('core/forge-process');
 
 async function isForgeStopped() {
   const pid = await findServicePid('forge_starter');
