+++ src/cli/node/join/join.js
@@ -9,7 +9,7 @@ const get = require('lodash/get');
 const set = require('lodash/set');
 const GraphQLClient = require('@arcblock/graphql-client');
 const { config, ensureConfigComment } = require('core/env');
-const { findServicePid } = require('forge-process');
+const { findServicePid } = require('core/forge-process');
 const { symbols } = require('core/ui');
 
 async function main({ args: [endpoint = ''], opts: { yes } }) {
